python3 migration

+++ gif.py
@@ -13,7 +13,7 @@ def search_gif(query, limit=25, offset=0):
 
 def random_gif_url(query):
     total_gifs = search_gif(query=query)
-    random_offset = randint(0, (total_gifs['pagination']['total_count'] - 1) / 25)
+    random_offset = randint(0, int((total_gifs['pagination']['total_count'] - 1) / 25))
     gif = search_gif(query=query, limit=25, offset=random_offset)
     random_index = randint(0, gif['pagination']['count'])
     return gif['data'][random_index]['images']['fixed_width']['url']
